위치설정 오류 수정, 참여중인 캠페인 리스트 확인 ui 추가 (이게 진짜)
