+++ src/atoms/elements/texts.tsx
@@ -46,6 +46,7 @@ export const WhiteText = styled.Text`
     color: white;
 `
 
+
 export const TitleBadge = (props: ITitleBadge) => {
     return (
         <Text style={{
