+++ src/navigation/GameNav.tsx
@@ -2,6 +2,7 @@ import React from 'react'
 import { createStackNavigator } from '@react-navigation/stack';
 import { colorCode, HeaderLeftBackIcon } from '../atoms';
 import GamePlayStack from '../container/GameNav/GamePlayStack';
+import QuizStack from '../container/GameNav/QuizStack';
 
 const Stack = createStackNavigator();
 
@@ -20,6 +21,13 @@ const GameNav = () => {
                 component={GamePlayStack}
                 options={{ headerTitle: "게임 플레이" }}
             />
+
+            <Stack.Screen
+                name="QuizStack"
+                component={QuizStack}
+                options={{ headerTitle: "퀴즈 플레이" }}
+            />
+
 
         </Stack.Navigator>
     )
